Faster caching of prototype accesses.
Summary:
We have a kind of "prototype caching" for properties found on an
object's prototype chain, but it is only used after a dictionary
lookup ensures that the property is not an own property of the object,
which decreases it's effectiveness as an optimization.

This diff changes prototype caching.  A PropertyCache entry gains a
new field: the "negMatchClazz".  When we find a property in the
(immediate) prototype (and therefore not in the object), we set the
"clazz" field of the PropertyCache to the HC of the prototype object,
and the "slot" field to the slot index of the property in the
prototype.  This is not sufficient, however: between one access and
the next, the property might get added to the object, and then it
would then be incorrect to retrieve it from the prototype.  To guard
against this, we record the HC of the object in "negMatchClazz" --
when the object had this HC, the property was not in the object.  Both
HC's must match to get a prototype cache hit.  But when they do, we
avoid doing a dictionary lookup in the object.

Follow-ons planned:

Implement this optimization for shermes, and for
the JIT.  The structure of the optimization is such that things still
work correctly, but they may be slightly slower.

Currently, the prototype cache path checks that the current object is
not lazy, a proxy, or a host object.  The latter two, at least, would
be unnecessary if we ensured that proxy and host objects had an HC
distinct from that of an empty object.  Hopefully, something similar
would apply for lazy objects.

Reviewed By: neildhar

Differential Revision: D65297624

fbshipit-source-id: a7dce0e17944d11b75950306eb1666894e43665b