Fix assert when throwing a Proxy
Summary:
Fix assertion fired when throwing a Proxy. We're removing the assert
because having a function that cannot be used when it's a particular
type of object is confusing.

Additionally, if the assert is removed, then we'd call `getParent` on
JSProxy. That means the JSProxy's prototype becomes observable. JSProxy
currently uses objectPrototype even though the prototype should just be
null. Since it'll become observable, we need to fix it properly now.

Reviewed By: neildhar

Differential Revision: D68246102

fbshipit-source-id: f99425fd4c441fb7236cf5a185a30564b60278c5