Add `StringEncodingType` for String/PropNameId encoding in SynthTrace Records
Summary:
Adds an enum `StringEncodingType` to describe the encoding type of a
String or PropNameID when it is created. `CreatePropNameIDRecord`
already does something like this. CreateStringRecord` relies on a flag
to indicate if it's ASCII or UTF-8, which makes it hard to add UTF-16 as
well.

Next diff in stack will add UTF-16 support. This diff is just a small
refactor.

Reviewed By: dannysu

Differential Revision: D67303131

fbshipit-source-id: f84b200b1f5cff9d2c28257d4122532ee525a8ac