Fix Unicode property escape inverted bug
Summary:
Fix an edge case in the inversion logic of `addRangeArrayPoolToBracket`.
In order to perform an inversion, this function adds the complement
ranges of the input ranges. For example, to invert a single range [A, B]
we add two ranges: [0, A-1] & [B+1, UNICODE_MAX_VALUE]. There is a risk
when adding in the second range though. If it's the case that we are
inverting a range which already sits at the end of the unicode space,
then there is no remaining opposite range to add. This condition is
currently unchecked, which results in underflow and ruins the rest of
the computation.

Add in this check, and add a LIT test to make sure we compile this
correctly.

Reviewed By: avp

Differential Revision: D68233638

fbshipit-source-id: f00a136675981ada01ccec7726969ac598d58543