Share FileAndSourceMapIdCache between lazy compilation invocations
Summary:
It is possible for the `sourceMappingURL` to be tens of MB,
resulting in slow lookups, which is why this cache exists.
It's not reused between lazy compilations when it's instantiated
inside `BytecodeModuleGenerator`, so allow it to be passed in
and store it in `BCProviderFromSrc` for lazy compilation purposes.

This wasn't a problem in Hermes because the `stripSourceMappingURL`
option was being provided to the lazy compilation invocations,
but that seemed more fragile than just doing everything the same
and reusing the cache for performance optimization.

This is made easy enough by moving the definitions into their own
header to avoid circular dependencies in the HBC directory.

Reviewed By: neildhar

Differential Revision: D68166926

fbshipit-source-id: 14f52273f4642df91dca2db0ca2011fa11b9a4f6