Fix Unicode property escape inverted bug
Summary:
Original Author: fbmal7@meta.com
Original Git: 159752fb95c6b4fb84048d064b94ca980b1054b6
Original Reviewed By: avp
Original Revision: D68233638

Fix an edge case in the inversion logic of `addRangeArrayPoolToBracket`.
In order to perform an inversion, this function adds the complement
ranges of the input ranges. For example, to invert a single range [A, B]
we add two ranges: [0, A-1] & [B+1, UNICODE_MAX_VALUE]. There is a risk
when adding in the second range though. If it's the case that we are
inverting a range which already sits at the end of the unicode space,
then there is no remaining opposite range to add. This condition is
currently unchecked, which results in underflow and ruins the rest of
the computation.

Add in this check, and add a LIT test to make sure we compile this
correctly.

Reviewed By: avp

Differential Revision: D68636157

fbshipit-source-id: 6e85e3dc45f17449957c00ba266e4f701680c78d