Make BoundFunction non-lazy
Summary:
We currently create `BoundFunction` as lazy if the target is lazy, and
populate its properties from the target when they are accessed. However,
this doesn't account for the possibility that the target is modified
after the `BoundFunction` is created. If that happens, looking up the
property on the `BoundFunction` will retrieve the new value instead of
the value that was present when the function was bound.

To address this, make `BoundFunction` always non-lazy.

This should have very little effect in practice, since looking up
`target.bind` will itself make `target` non-lazy, so most
`BoundFunction`s should already be non-lazy.

Reviewed By: avp

Differential Revision: D68638177

fbshipit-source-id: 50726e819544efe2590f0ff7c79df9484b77c0b8