Update Native backend with proto cache.
Summary:
Previous work (stack ending in D65297624) introduced better caching of
property loads satisfied from an object's prototype.  That
optimization was done only in the interpreter -- the structure was
such that the native backend and JIT would continue working.

This diff extends the optimization to the native backend.  The JIT
generates inline code to check for "own property" cache hits, then
falls back to the same function the native backend uses, so it also
gets the benefit of this optimization.  We could consider adding
inline code for the proto cache, but the cost/benefit tradeoff is not
clear.  However, we probably should make versions of functions like
_sh_ljs_get_by_id_rjs that assume they're called only after a failed
cache check, and don't repeat that work; I'll make a task for that.

Reviewed By: neildhar

Differential Revision: D68515127

fbshipit-source-id: 9c9499b9e99d3091631664b6bc535e0ffa2ecaa9