Use the same AndroidUnicodeUtils.java everywhere
Summary:
hermes and hermes-snapshot have this issue where including both and use
their own copies of AndroidUnicodeUtils would cause a package name
clash. This is because by having different dependencies, BUCK will try
to include 2 packages that have the same package name. The solution
used before was to have both take dependency on the same copy of
AndroidUnicodeUtils.

Now that we're trying to include both Hermes and Static Hermes, we have
the same exact issue. For now I'm taking the same approach by moving to
a common copy. The code is identical so there isn't a behavioral change.

Reviewed By: neildhar

Differential Revision: D68173770

fbshipit-source-id: ae62a35a49fc7b8b9680cb47f40457137f717c4d