Revert "Disable DateTimeFormat::formatToParts for Apple platform (#1155)" (#1567) (#1567)
Summary:
Original Author: ncor@meta.com
Original Git: ac557bb00fa815946c87732c7b81174a73cde5be
Original Reviewed By: blakef
Original Revision: D66099176

This reverts commit c5a633f6452902cd0b5c9392a1e4f51a33933284.

That commit removed the Intl.DateTimeFormat.formatToParts for Apple platforms from Hermes.
However, the community started depending on it, and we ended up manually reverting that commit in the release branch for the past 3 releases.

This is like reverting the commit in main, so we are proceeding with that approach to simplify the release process of React Native and Hermes to the OSS.

Pull Request resolved: https://github.com/facebook/hermes/pull/1567

Pulled By: cipolleschi

Reviewed By: neildhar

Differential Revision: D68868242

fbshipit-source-id: c06b9821feb57507c7380358a7181747b91b0de5