Consolidate all compiler calls from VM into HBC.h
Summary:
Reduce the amount of classes imported into the VM from BCGen,
allowing us to more easily swap out the actual compiler for stubs.
This will let us remove the `HERMESVM_LEAN` macro and avoid recompiling
the entire VM twice.

Make a couple of `BCProviderFromSrc` functions virtual and avoid
using `BCProviderFromSrc` directly in the VM.
It's not able to be made entirely private because other places still use
it, but all the VM and API users will go through HBC.h only,
using `BCProvider` as the only imported class.
This lets us keep the lean stubbed-out functionality in a single place.

This does mean that we allow HBC.h to take in `BCProvider` instead of
only `BCProviderFromSrc` to avoid having to pull it in.

Users now have a simpler API for `createBCProviderFromSrc` in HBC.h.

Reviewed By: neildhar

Differential Revision: D68842567

fbshipit-source-id: 426462bd712e0118602236d1bef327dbffe8c805