Cache the object's HC for use in GetNextPName
Summary:
The first step of speeding up `for-in` iteration over objects
is to avoid doing unnecessary membership checks every iteration.
This is done by also caching the object's own HiddenClass
(we're already storing the rest of the prototype chain)
along with the number of properties it has.

Then on each iteration we can do some simple math and check that the
HiddenClass hasn't changed instead of running `findProperty`.

Reviewed By: tmikov

Differential Revision: D67292849

fbshipit-source-id: a1807d0d1bb5ef1a6bf318bb32e8088ab2bc578b