Gracefully return if dlopen/dlsym has error
Summary:
In case like ICU lib path is not set on Linux, dlopen will fail and
return NULL, which causes dlsym to return NULL and ASAN reports an
error when accessing that. On some platforms, NULL is interpreted as
special handle (e.g., RDTL_DEFAULT) by dlsym, and a random function is
returned. We should check `dlerror()` after each dlopen/dlsym call and
gracefully exit if there's an error.

Reviewed By: neildhar

Differential Revision: D67215415

fbshipit-source-id: 1f5eb0871ed967102a16c928e3747458821d53d2