Parse source map in BCProviderFromSrc::create
Summary:
We currently fail to link `hermesvmlean` because we take a dependency on
the `SourceMapParser` which in turn requires the `JSONParser`. We
currently only use it to pass a parsed `SourceMap` to
`createBCProviderFromSrc`, so instead pass the actual source map and
perform the parsing there. This allows us to compile out the
`SourceMapParser` in lean builds.

While we're changing the signatures, clarify and enforce the null
termination requirement for the source map.

Reviewed By: avp

Differential Revision: D68936292

fbshipit-source-id: c1d00630c85c7cdb4a461100abd62cc7aa41f3bf