Skip tracking users for literals
Summary:
The user lists for literals can grow extremely large since they are
shared across the module, and may occur many times in object and array
literals. Updating usage of literals can be very costly, since it needs
to iterate the user list to remove the prior use.

One example where this is currently very slow is when deleting
AllocObjectLiteral when it gets lowered:

1. Iterate over the (potentially large) list of operands of the
instruction to remove each one.
2. For each operand, iterate the user list of the corresponding literal
to find its use.
3. Once the use is found and removed, the last user gets moved into its
position. This requires updating the corresponding `Use`, which means
iterating through the operands of the using instruction.

Reviewed By: avp

Differential Revision: D67290808

fbshipit-source-id: 9344fecc5d56ebdeab25a821c00dc299aed46cda