Identify profiler suspend reasons (#1608)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1608

The sampling profiler tracks the type of each frame. A "suspend" frame
denotes that execution is suspended, with an arbitrary string providing
details.

We would like to identify GC frames for the sake of Chrome DevTools
visualizations/stats about GC time. This diff adds an enum to suspend
frames to allow GC frames to be identified without depending on
specific contents of the details string.

Another option would be to replace the "suspend" frame with multiple
different frame types for debugger, GC, etc.

Reviewed By: fbmal7

Differential Revision: D68442991

fbshipit-source-id: d16a3a1b87657e538e8c5471ea9533876f1c757d