Refactor CodeMotion to run on instructions instead of operands
Summary:
`CodeMotion` currently works by looking at the operand of each
instruction and then sinking that operand to the location of the
instruction if it is eligible. Refactor it to instead consider the
instruction it is currently looking at for sinking.

This allows us to add the ability to sink instructions with multiple
users in upcoming diffs.

Reviewed By: avp

Differential Revision: D61550416

fbshipit-source-id: 27080a5fa5787ebd223aebba5c364e536ae55a62