Fix JIT implementation of CoerceThisNS
Summary:
Address a few issues in the implementation of CoerceThisNS.

We currently do not always sync the `frThis` operand to `CoerceThisNS`
to the frame, even though it needs to be read in the slow path. This is
both because we don't check `frThis == frRes` when deciding whether to
skip syncing `frRes`, and because we do not force a sync of `frThis` to
the frame.

We also allocate the temporary register `hwTemp` after freeing `hwThis`,
which creates the possibility of the two being assigned to the same
register.

Reviewed By: avp

Differential Revision: D69508167

fbshipit-source-id: e3b8f57e71c414e5b1f38356ab141d3a3f4ea23d