Allow tracing through child scopes in FunctionAnalysis
Summary:
Allow `FunctionAnalysis` to emit `ResolveScopeInst`s if the scope that
is available at the point of a call is a descendent of the scope we
actually want. This will allow us to trace a function through inner
scopes while retainin the ability to emit direct calls, or inling
without having to emit a `GetClosureScopeInst`.

Reviewed By: avp

Differential Revision: D68187126

fbshipit-source-id: 12a6132e7daf99bfd1423f0144d263bfa108f0e2