Propagate more scopes in FunctionAnalysis
Summary:
Detect when a closure is being stored into an inner scope, and use that
to propagate the scope, since the closure scope is still reachable from
the inner scope. This is especially useful in the presence of scope
hoisting, because the closure's parent may be changed to some outer
scope, while it is still being stored in the original scope.

Reviewed By: avp

Differential Revision: D68187124

fbshipit-source-id: 6ed0f408cba5c7cc5fcef9a9fed234ad19f32fc6