Increase timer resolution for sampling profiler
Summary:
When using the sampling profiler on Windows, it is apparent that the
frequency is not as high as it is on other platforms. Doing some local
testing, the cause of the low frequency sampling is Windows timer
resolution when calling std::condition_variable::wait_for. By default,
this resolution is approximately 64Hz, meaning the sampling profiler
maxes out at ~64Hz. Setting `timeBeginPeriod(1)` improves sampling
frequency, maxing out at ~500Hz.

Reviewed By: neildhar

Differential Revision: D69500253

fbshipit-source-id: 3108c4629e6f2853731171e92ddf6af89792154b