Align HermesRuntime with SH and make non-static methods virtual
Summary:
Make these functions virtual so that it is simpler to have multiple
different Hermes versions loaded in the same application.

This should not add any overhead in most cases, since Hermes is
typically a separate dynamic library, so invocations of these methods
would take an indirection through the PLT.

Also add methods that currently are implemented in SH, with throwing
implementations.

Reviewed By: tmikov

Differential Revision: D69219627

fbshipit-source-id: 6eabd012d05ff1c529750781a5f560881e2fc700