Track number of frames sampled by SamplingProfiler directly in the sampled vector
Summary:
Currently, we allocate a vector of size 500 to store the stack frames in
the Sampling Profiler. The variables `sampledStackDepth_ ` and
`preSuspendStackDepth_` are used to track how many frames are actually
collected during a sampling event.

This refactor makes it so that the size of the vector reflects the
number of collected frames, rather than tracking it in the extra
variables.

Reviewed By: neildhar

Differential Revision: D69201808

fbshipit-source-id: 1007b6e395e65521b5648038969af0a605a2114f