Save the thread ID in SamplingProfiler
Summary:
The sampling profiler currently records the thread ID during stack
walking. This works fine on POSIX platforms because the signal handler
actually runs on the runtime thread, but causes problems on Windows,
where it occurs on the timer thread.

To fix this, record the thread ID when the runtime is registered for a
particular thread, and just use that when walking the stack.

Reviewed By: rozele

Differential Revision: D69664487

fbshipit-source-id: d6d1d46729709505a563efbd498f9bc253ba2ef8