Refactor TypeOfIs in the JIT to depend less on HV layout
Summary:
Instead of creating the HV upfront and then setting the low bit, compute
the low bit first and then set the HV tag at the end. This introduces an
extra move of 0 or 1 in some cases, but it removes the dependency on HV
layout, which is useful for the layout changes in upcoming diffs.

Reviewed By: avp

Differential Revision: D69795898

fbshipit-source-id: 9ea0fe5c26da1fab45207e3bcf4bbf7714b7f808