Add scope hoisting
Summary:
Add a pass that analyses all of the scopes that are transitively
resolved by a given `VariableScope`. It then hoists the scope past any
scopes that it does not "use".

This allows us to flatten the tree of scopes, which in turn means that
each scope retains less memory at runtime, and has to traverse fewer
parents during scope resolution.

Reviewed By: avp

Differential Revision: D60728338

fbshipit-source-id: a743981e8baecbe703197f546cfdc354c4d724a9