Hoist functions to outer scopes
Summary:
Now that we hoist the scopes inside functions, we can also hoist a
function such that its parent is the nearest scope that it actually
uses.

This reduces the amount of scope resolution that needs to take place
on each call to the function, since it happens once at creation.
Importantly, it allows some functions to be hoisted to the top level,
such that they do not have a scope at all.

As with hoisting scopes, this also reduces memory consumption, since a
function will now avoid retaining scopes that it does not have to.

Reviewed By: avp

Differential Revision: D60748882

fbshipit-source-id: 683fcda6ea08127557904eec05a5e978218da782