Support/BitField: utility class for bit-field access
Summary:
Add a convenience class to access bit fields, when the layout of the
fields must be guaranteed and compiler independent (e.g. for on-disk
storage).

Note that it only guarantees the bit offsets within the larger
storage. The alignment and the "endianness" must be ensured in different
ways, usually by aligning fields in a struct to "natural" sizes and
requiting little-endian in our case.

`BitField<StorageType, FieldType, BitPosition, Width>` is an utility
type that provides static accessors to validate, extract or update a
bit-field in a larger storage.

Besides the accessors, it provides a type alias for a `NextField`, which
knows the storage type and bit position.

Inspired by v8's BitField, but should handle signed types correctly.
An LLM wrote the first version, but it was both buggy and sub-optimal,
so some biological intelligence was involved.

The usage pattern is to define a type alias and accessors for every
field. For example:
```c++
// Before
struct Flags {
  uint32_t a : 10;
  uint32_t b : 15;
};

// After
struct Flags {
  uint32_t _data;
  using AField = BitField<uint32_t, uint32_t, 0, 10>;
  using BField = AField::NextField<uint32_t, 15>;

  uint32_t getA() const { return AField::extract(_data); }
  void setA(uint32_t x) { _data = AField::update(_data, x); }
  // And so on...
};
```

Reviewed By: avp

Differential Revision: D70069551

fbshipit-source-id: 630e1f7573f16feeb87766cf08d353220460a81c