Add dependency to ensure library order in linking.
Summary:
In other work, I introduced a call in an optimization pass to a
function defined in the IR directory.  This built fine on Mac.  But I
was attempting to build on Linux; clang there seems to be more picky
on library load order, and gave an unresolved symbol error because
libOptimizer.a came before libhermesFrontend.a.

This diff alters the CMakeList.txt file makes the hermesOptimizer
library depend on hermesFrontend.  That solves the problem on Linux.

Reviewed By: neildhar

Differential Revision: D70413618

fbshipit-source-id: 49fa3f715ad3e06c5741a162126539dcd2a8500e