Distinguish JSSpecialParam from JSDynamicParam.
Summary:
The class JSDynamicParam has a method getIndexInParamList(), which
returns the index of the JSDynamicParam in its owning function's param
list.  This asserts via an llvm::unreachable if the param isn't found.

But today there are some instances of JSDynamicParam that are not in a
function's param list: the newTargetParam_ and parentScopeParam_
fields of a Function.  So it isn't legal to call getIndexInParamList()
on these.

After some discussion, we hit on the solution in this diff: separate
these into two kinds/classes.  I chose JSSpecialParam for the kind
that isn't in the param list.  It worked out fine to segregate things
this way; there's never any ambiguity about which kind a param is.

In the discussion, I proposed that JSSpecialParam could be a supertype
of JSDynamicParam; the latter would add only the getIndexInParamList()
method.  I did them completely separately at first, and thought that
seemed like a little too much code duplication.  So I made a common,
implementation-only supertype, JSParam, to capture the common stuff.
This has a protected constructor, so it's only there as an
implementation technique.

Reviewed By: fbmal7

Differential Revision: D60611614

fbshipit-source-id: 6880b49dcedd1b40fb79cea4d97f6664ed1dde90