Fix CreateThis handling in FunctionAnalysis
Summary:
`FunctionAnalysis` currently only accounts for the use of closures as
the callee argument to `CreateThis`, rather than the `new.target`. If
these are not the same, it can result in an assert.

This means that simple programs like:
```
function foo(){
  class A{}
  class B extends A{}
  return new B();
}
```
cause a crash in the compiler if the inlining threshold is raised.

Split the handling of these two, since the callee determines the result
but the new target determines whether the closure has leaked.

Reviewed By: fbmal7

Differential Revision: D70542803

fbshipit-source-id: 509881b8c6f071a925889dbcec7422e70fd28442