fix stepping into blackboxed range with manual breakpoint
Summary:
In themost use-cases, manual breakpoints are triggering a pause from `handleDebuggerEvent` with the event `DebuggerEventType::Breakpoint`.
However, when we step into a blackboxed range and keep on stepping into inside of it, we need to explicitly check if that line has a manual breakpoint, because it does not trigger `handleDebuggerEvent`.

Reviewed By: dannysu

Differential Revision: D70617441

fbshipit-source-id: 18bff9e1f5a2800737e7542e0c39479c10684f32