Handle Phi users in ObjectMergeNewStores
Summary:
`ObjectMergeNewStores` depends on the invariant that the
`AllocObjectLiteralInst` dominates all of the users we identify.
However, this isn't true of Phis, which may not be dominated by the
instruction and therefore may be missed in the dom tree traversal. I
don't believe this can currently cause bugs, but there is nothing
structurally in the IR preventing it.

This can also result in suboptimal code if the Phi occurs in the same
block as the `AllocObjectLiteralInst`, because it appears as though the
object immediately escapes.

To address these issues, add explicit handling of Phis.

Reviewed By: avp

Differential Revision: D70500582

fbshipit-source-id: d457823ebd463ce4a210636e29a620a5e78cb7bb