Move recording of individual collections outside the GC
Summary:
The ShouldRecordStats parameter on GCConfig mostly records things that
already can be done directly by the integrator, but without fine grain
control over the amount of information that gets dumped.

One particularly egregious case is that the stats for every collection
are always saved and reported when it is enabled. This means that the
GC has to record a potentially unbounded amount of information when the
flag is specified. It also makes it hard to discard this data if we are
only interested in the top-level metrics.

To address this, move the recording and reporting of per-collection
stats outside the GC.

Reviewed By: dannysu

Differential Revision: D70741975

fbshipit-source-id: 88fbbdd5e8b9720ebb3a3c457778c8da957a8b8d