Move marking state into HadesGC
Summary:
The state for an ongoing OG marking phase is currently stored in a heap
allocated `MarkAcceptor`. Instead, split the state so that persistent
fields are stored inline in `HadesGC`, while things like `GC&` and
`PointerBase&` are kept in `MarkAcceptor`, which is now lightweight
enough to be instantiated on the stack as needed.

This has a few advantages:

1. There are fewer indirections for accessing the persistent state,
since it is inline in the GC
2. It is easier to template the OG acceptor, since it is no longer a
field in the GC, and can instead be instantiated when needed (like the
YG acceptor)
3. The "transient" data is on the stack, and can be promoted to
registers with sufficient inlining of the marking code. This is
potentially more efficient.

Reviewed By: lavenzg

Differential Revision: D70862800

fbshipit-source-id: d010d1e5a679c7dbe2bcd16b737b06f4da2abeae