Copy InternalBytecode to the heap when debugger is enabled
Summary:
The debugger depends on being able to breakpoint any `CodeBlock`,
but when the bytecode lives in rodata in the binary the `mprotect` call
fails. There's potentially ways to solve this that allow failing to set
breakpoints or avoiding trying to set them but they come with some
complexity.

To fix the breakage for now, copy the data into the heap when the
`Runtime` is constructed and run off that instead.

Reviewed By: neildhar

Differential Revision: D71163360

fbshipit-source-id: 9e85e8cbe666c8ec800e7441c92d31ea42403192