Add a new PrivateName IR Type
Summary:
This type is what we will use to generate the value of a private name.
We can then use this type in IRVerifier to make sure all the new private
loads/stores only operate on this type.

Corresponds to: https://tc39.es/ecma262/2024/#sec-private-names

At runtime, this type will be a symbol, but it doesn't *have* to be
implemented that way, and conceptually they are different entities.

Reviewed By: avp

Differential Revision: D69217588

fbshipit-source-id: 71b0adb21ed19443882e5f1c8cf2853e40fbab67