Validate private names
Summary:
Enforce the rules for `PrivateBoundIdentifiers` described in
https://tc39.es/ecma262/#sec-class-definitions-static-semantics-early-errors.
Specifically the following:

> It is a Syntax Error if the PrivateBoundIdentifiers of
ClassElementList contains any duplicate entries, unless the name is used
once for a getter and once for a setter and in no other entries, and the
getter and setter are either both static or both non-static.

Also error on `super.#privateName`.

Reviewed By: avp

Differential Revision: D64562762

fbshipit-source-id: fa77982d6f6de3d27fdbfc9bf41d5d2abcb60b1a