Mark CacheNewObject key operands as identifiers
Summary:
Key operands to `CacheNewObject` are identifiers. This was previously
hidden because most keys are also string operands to some other property
store operation. However, if the key is a number, it may be normalised
to a `LiteralNumber` on the store operation, while remaining a
`LiteralString` on the `CacheNewObjectInst`, triggering this bug.

Reviewed By: avp

Differential Revision: D71404303

fbshipit-source-id: ba0ea85efe7ed50795351d7bfa6bb80d485af793