Explicitly populate direct property storage of JSObject
Summary:
On Apple platforms, the `uninitialized_fill` loop gets turned into a
call to `memset`. This is considerably slower than just populating the
direct property slots directly. To prevent this, emit a sequence of
separate stores, which avoids emitting the call.

On the simple benchmark below, this reduces the time on my M4 Mac from
~700ms to ~580ms.

```
(function (){
  let t1 = Date.now();
  for (let i = 0; i < 100_000_000; ++i){
    globalThis.x = {};
  }
  let t2 = Date.now();
  print(t2 - t1, "ms");
})();
```

Reviewed By: avp

Differential Revision: D71156203

fbshipit-source-id: ddc43243ed6a2cf00fad62be8f15f1000dafeb8f