Delete Promise VM option and hasAsync bytecode flag
Summary:
`-Xes6-promise` was a hidden experimental flag that has now run its
course. It's defaulted to on, so delete it now that Promise is always
enabled. `hasAsync` was a bytecode option that was only ever used to
check against the Promise flag on the VM, and now that the VM
configuration is gone, we can also remove the hasAsync flag from the
bytecode.

This diff does NOT delete `HermesInternal.hasPromise()`, just makes it
always return `true`, to avoid breaking any existing callers.
Also, it's a convenient read-only function and there's no reason to
delete it.

Reviewed By: fbmal7

Differential Revision: D71507961

fbshipit-source-id: 288ed716eb62a29579089d7a353463349a8cd1bc