Slightly reduce branching and simplify SerializedLiteralParser
Summary:
`SerializedLiteralParser` was iterating through both `totalLen` and
`leftInSeq` together in the inner loop on each tag, to allow stopping
either when the number of requested values is reached or when the
values under the current tag run out.

The number of values to read for each tag can be determined up front,
making it easy to just loop `toRead` number of times within in case in
the `switch`. This is very slightly faster but it's also easier to
understand.

Reviewed By: neildhar

Differential Revision: D71657086

fbshipit-source-id: d2b4afb8624f639d5ba2f1c1ca1cb28505650957