Avoid reordering "Other" registers
Summary:
These registers may be pointers, and they don't need to be reordered
within their class regardless of where they are in the function,
because reordering them doesn't help the JIT do register allocation like
it does when it puts numbers/non-pointers into hardware registers when
compiling, and when we do reorder them, it hurts the compressibility of
result bytecode.

So this mitigates the compressed size regression of
`-freorder-registers`.

Reviewed By: tmikov

Differential Revision: D71582891

fbshipit-source-id: 8e1ec0ff5f9bcaaff00458104c25a06f151fd9db