Change Utils.cpp to use a namespace; fix lint complaint.
Summary:
When doing other things, I noticed that Utils.cpp used a legal, but
non-standard, method of defining symbols in a namespace (namely, doing
"using namespace hermes;" in the .cpp file).

On the theory that a codebase should try to be consistent, this diff
changes that to use the more standard technique of entering the
relevant namespace in the .cpp file.

Reviewed By: avp

Differential Revision: D71591231

fbshipit-source-id: 755c01db96e86386e5b7e7ea8a63d366f19edccf