Implement StoreOwnPrivateFieldInst for JIT / native backend
Summary:
Implement storing to private fields in the native/JIT
backend. This consists of declaring a new C function in
`static_h.h` which is invoked from both the native backend
and the JIT handler for the BC instruction
`PutOwnPrivateBySym `. If the private field doesn't exist,
then an error is thrown. The operation is cached on both the
HC of the target object and the symbol value of the private
name.

With this diff, all private name operations are complete in
the JIT and native backends, so tests can be turned on for
the LIT tests that were added earlier in the stack.

Reviewed By: avp

Differential Revision: D71587531

fbshipit-source-id: b712821d045f7f9a4d8539ae822ee0d40d55b273