Implement LoadOwnPrivateFieldInst for JIT / native backend
Summary:
Implement loading private fields in the native/JIT backend.
This consists of declaring a new C function in `static_h.h`
which is invoked from both the native backend and the JIT
handler for the BC instruction `GetOwnPrivateBySym `. If the
private field doesn't exist, then an error is thrown. The
operation is cached on both the HC of the target object and
the symbol value of the private name.

Reviewed By: avp

Differential Revision: D71587527

fbshipit-source-id: 015e7e2eae3e4463fc59dd09595035f131bb5280