Implement AddOwnPrivateFieldInst for JIT / native backend
Summary:
Implement the add private field operation in the native/JIT
backend. This consists of declaring a new C function in
`static_h.h` which is invoked from both the native backend
and the JIT handler for the BC instruction
`AddOwnPrivateBySym `. The private name symbol key of the
field must not already exist on the object.

Reviewed By: avp

Differential Revision: D71587530

fbshipit-source-id: e23a1db055dfe51d270232e93f6986d7ae4df5f2