Create private name cache in native backend
Summary:
Add a new cache in the native backend which will hold
entries for all private name operations (read, write, `in`.)
This mirrors the HBC backend: each cache entry holds the HC,
WeakSymbolID, and a slot index. This cache must be scanned
through during `sh_unit_mark_long_lived_weak_roots`. And
`SH.cpp` is modified to support creating this cache with the
appropriate size and reserving slots in the cache for any
callsites that need a `SHPrivateNameCacheEntry` via
`genPrivateNameIC`.

Reviewed By: avp

Differential Revision: D71587532

fbshipit-source-id: b4d6d86f61bde946a5839717ea2bc28bdc366f85