Declare `arguments` for field initializer values
Summary:
The field initializer value of a class property conceptually belongs in
its own function. We are already modeling this concept in
SemanticResolver via the the new FunctionContext we make, paired with
`getOrCreate(Static|Instance)ElementsInitFunctionInfo`. This new
function that we are making should also declare it's own `arguments` as
a special identifier, so that we can later error out if there's any
usage of it.

This fixes a bug where value initializers of classes in the global scope
would not error out on `arguments` usage.

Reviewed By: avp

Differential Revision: D71831165

fbshipit-source-id: 741194f6f80e4e778a4a52a0385d104e4aac3335