Implement adding private fields
Summary:
Add support for private fields in IRGen. Ensure that
double-initialization of fields is detected and an error thrown. Private
fields must be added sequentially according to their lexical order. So,
it is impossible for the following class to leave an instance
initialized with `#b` but not `#a`.

```
class A {
  #a = foo();
  #b = 2;
}
```

Given that `#a` *must* be defined if `#b` is defined, it is sufficient
to test for double initialization by only checking to see if the first
private element has been defined. If not, then all the others won't be
defined either.

Reviewed By: avp

Differential Revision: D64867417

fbshipit-source-id: ad51978305df2e49b2017d43f3a6adb5c86cd108