add JSTypedArrayBase::polyReadMayAlloc()/NoAlloc()
Summary:
`JSTypedArrayBase::polyReadMayAlloc()` and
`JSTypedArrayBase::polyReadNoAlloc()` perform a polymorphic read of a
 typed array element returning a HermesValue.

`polyReadMayAlloc()` supports all kinds of typed arrays and consequently
may need to allocate memory for a Bigint.

`polyReadNoAlloc()` only works for the subclass of typed arrays
identified by the new method `isNoAllocTypedArray()` and never
allocates, so it is more suitable for the interpreter loop.

Use `polyReadMayAlloc()` in `typedArrayPrototypeAt()`, which previously
had a switch. This fixes a bug where bigint typed array values weren't
returned as Bigint.

Add a test for int64.

Differential Revision:
D72099909
***
tmp: polyReadNoAlloc()